+++ crates/module-system/module-implementations/sov-evm/tests/integration/helpers.rs
@@ -47,7 +47,7 @@ pub(crate) fn setup() -> (TestRunner<RT, S>, TestUser<S>, EvmAccount, EvmAccount
     ));
     let genesis_config =
         HighLevelOptimisticGenesisConfig::generate().add_accounts(vec![rollup_account.clone()]);
-    let evm_config = EvmGenesisConfig {
+    let mut evm_config = EvmGenesisConfig {
         accounts: vec![
             AccountData {
                 address: evm_account.address(),
@@ -64,11 +64,11 @@ pub(crate) fn setup() -> (TestRunner<RT, S>, TestUser<S>, EvmAccount, EvmAccount
                 nonce: 0,
             },
         ],
-        // SHANGHAI instead of LATEST
-        // https://github.com/Sovereign-Labs/sovereign-sdk/issues/912
-        hardforks: vec![(0, SpecId::SHANGHAI)].into_iter().collect(),
         ..Default::default()
     };
+    // SHANGHAI instead of LATEST
+    // https://github.com/Sovereign-Labs/sovereign-sdk/issues/912
+    evm_config.chain_spec.hardforks = vec![(0, SpecId::SHANGHAI)];
     let genesis = GenesisConfig::from_minimal_config(genesis_config.into(), evm_config);
     let runner =
         TestRunner::new_with_genesis(genesis.into_genesis_params(), TestRuntime::default());